AC4 - Refactor to be able to swap startegy on how to format the results
