+++ frontend/src/search.js
@@ -4,6 +4,7 @@ import { refreshPaginationButtons,openPaginatePage } from './paginate/paginateBu
 import { ResultsPrinter } from './resultsContainer/resultsContainer.js'
 import { normaliseAndLowecase,splitTextInWords } from './helpers.js'
 import { clearAndShowSearchMessage,manageExceptionUI } from './general.js'
+import ShortenedLinesFormatter from './results/contentFormatters/ShortenedLinesFormatter.js'
 
 const MAX_RESULTS_IN_PAGE = 15;
 export async function processSearch(query,resultContainer,paginateButtonsContainer,HulipaaOpt) {
@@ -29,7 +30,8 @@ export async function processSearch(query,resultContainer,paginateButtonsContain
         return;
     }
 
-    const resultsPrinter = new ResultsPrinter(searchedWords,resultContainer,paginateButtonsContainer,HulipaaOpt)
+    const resultFormatter = new ShortenedLinesFormatter()
+    const resultsPrinter = new ResultsPrinter(searchedWords,resultContainer,paginateButtonsContainer,resultFormatter,HulipaaOpt)
 
     // This needs to happen before openPaginatePage, otherwise when it tries to highlight a button it can't find it
     refreshPaginationButtons(paginateButtonsContainer,finalResults,resultsPrinter,MAX_RESULTS_IN_PAGE,resultContainer)
